OPENDJ-1602 (CR-5566) New pluggable storage based backend
Add Persistit backend to the configuration framework:

* add XML config definition
* add config LDAP schema (OID needs reserving)
* switch pluggable backend to use new configuration
* removed ConfigurableEnvironment since it is very tied to JE.

At the moment the pluggable backend implementation is very tightly coupled to the persistit storage so it is not possible to define an abstract pluggable backend configuration. In particular, the pluggable backend assumes that the underlying storage is disk based and uses only a single directory for its data.


git-svn-id: https://svn.forgerock.org/opendj/branches/opendj3-server-dev@11478 41b1ffd8-f28e-4786-ab96-9950f0a78031