OPENDJ-1713: Persistit: dsconfig configuration properties have no effect
Hook configuration parameters into Persistit so that it is possible to configure the cache utilization and commit policy. It is a bit brittle, but seems to work on my laptop.

The commit policy config may need adjusting: JE provided three settings, effectively NONE, FLUSH, and FSYNC. Persistit has only two settings: FSYNC and non-FSYNC. It's unclear at this stage what the non-FSYNC mode is, i.e. whether it is flushing or not. I've adjusted the default commit policy to be SOFT otherwise full FSYNC is horribly slow.


git-svn-id: https://svn.forgerock.org/opendj/branches/opendj3-server-dev@11535 41b1ffd8-f28e-4786-ab96-9950f0a78031