(CR-5887) Decouple LDIFReader from JE's import
Currently LDIFReader depends on JE's importLDIF classes.
This will become a problem when trying to build JE as a separate jar from OpenDJ zip.
This is already a problem with pluggable backends because it means we would need to make LDIFReader also depend on the pluggable import package, which is not acceptable.


Extracted ImportLDIFReader as a subclass of LDIFReader to receive all import related code + used the former in Importer class.


git-svn-id: https://svn.forgerock.org/opendj/branches/opendj3-server-dev@11670 41b1ffd8-f28e-4786-ab96-9950f0a78031