CR-6060 OPENDJ-1785 Add "extra" doc into code for generated content
This patch adds supplemental documentation
for what is already part of the hand-written reference
to the generated reference content.
For example, descriptions that are longer in the hand-written reference
than in the generated reference content
are now included in the generated reference content.

This patch does not, however, add additional reference sections
that are not currently part of generated content,
such as EXIT CODES, EXAMPLES, SEE ALSO, etc.

This is a subtask of the following issue:
OPENDJ-386 Move to single reference document

git-svn-id: https://svn.forgerock.org/opendj/trunk@11722 41b1ffd8-f28e-4786-ab96-9950f0a78031