CR-6068 OPENDJ-1785 Add "extra" doc into code for generated content
This patch is like r11722, but for the LDAP SDK tools.

This patch does not resolve the fact that duplicate tools exist.
I have, however, logged OPENDJ-1810 regarding that point.
Duplication should disappear in the move to a single set of tools.

This is a subtask of the following issue:
OPENDJ-386 Move to single reference document

git-svn-id: https://svn.forgerock.org/opendj/trunk@11735 41b1ffd8-f28e-4786-ab96-9950f0a78031