CR-6565 OPENDJ-1806 Workaround usage() formatting in docs
Whitespace is not generally significant in doc sources
except where markup specifies that it is
such as within ProgramListing, Screen, and LiteralLayout elements.

Some usage messages in the generated reference docs
are formatted for use in terminal windows,
with significant whitespace and newlines
that get lost when pulled as is into the documentation source.

This patch works around usage() message content
that contains significant whitespace, specifically newlines.

git-svn-id: https://svn.forgerock.org/opendj/trunk@12023 41b1ffd8-f28e-4786-ab96-9950f0a78031