CR-6570 OPENDJ-1898 Move to single doc gen plugin
This patch brings the doc generation plugins together
in a single module, and adapts the implementation
of the log ref generation plugin to use a FreeMarker template
as is already in use in other plugins.

This patch also uses sections in the log ref.
Sections, rather than naked variablelists,
make the log ref easier to navigate.

git-svn-id: https://svn.forgerock.org/opendj/trunk@12027 41b1ffd8-f28e-4786-ab96-9950f0a78031